share test code
